workaroud winograd format issue in inference
If prop_kind == forward_inference, the mkldnn_wino_fmt for weights is required by winograd primitive.
Then, in the cases of variable input shape, the detials of mkldnn_wino_fmt will be changed.
And, extra weihgts reorder is inevitable each time, leading to bad performance.
Here, we set the prop_kind to forward, in order to reorder and cache weights as blocked format,
instead of mkldnn_wino_fmt.

Signed-off-by: Gu, Jinghui <jinghui.gu@intel.com>